requirements.txt
line-bot-sdk==1.20.0
gunicorn==20.1.0
Flask==2.0.2
gspread==4.0.1
oauth2client==4.1.3